Extract `CategoryManagement` into a standalone component and integrate it into `AppSidebar`. Simplify `AppSidebar` by delegating category management responsibilities.
